// src/app/models/movie/movie-getmoviebyid-response-data.interface.ts
export interface GetMovieByIdResponseData {
  "id": string;
  "title": string;
  "directorName": string;
  "genre": string;
  "contents": string;
  "runningTime": number;
  "posterUrl": string;
  "stillUrl": string;
  "favorite": number;
  "nation": string;
  "company": string;
  "ratedYn": boolean;
  "type": string;
  "actor": string;
  "releasedAt": string;
  "createdAt": string;
  "modifiedAt": string;
}